회ìfeat : 회원가입, ì유저 닉네임 중ë³ 아이디 중복체크 기능 구현

+++ gabojago_be/src/main/java/Gabojago/gabojago_be/entity/ExchangeRate.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 public class ExchangeRate {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int exchangeRateId;
+    private Long exchangeRateId;
     private String country;
-    private double rate;
+    private Double rate;
 }
